fix: init dns resolvers in stream subsystem.
Since f914807, the connect of tcpsock was patched by using our own DNS
resolvers. It relies on a dns_resolvers variable which should be
initialized in init_by_lua* phase, however, the initialization step in
stream sub-system is missed so once users enable the stream proxy and
the tcpsock:connect is invoked with a non-IP host, Lua runtime errors
will be thrown.

This PR fixed this issue by initializing the DNS resolvers in the stream
sub-subsystem.