ci: enforce request review on the final state
In https://github.com/apache/apisix/pull/5940#discussion_r802593858,
during solving the merge conflict,
a bug which is fixed in https://github.com/apache/apisix/pull/6204 is
reintroduced.
The test is also changed so it can't catch up the regression.

That PR is approved before merging the master. Luckily, I am cautious
enough to start another turn of review.

We should not rely on the lucky or someone else's caution.
To prevent that from happening again, here we enforce the approval
should be applied with the final state.

Signed-off-by: spacewander <spacewanderlzx@gmail.com>