chore: tweak conf_server unstable test cases (#7498)
### Description

In tests 6&7 of `t/deployment/conf_server.t`, two separate instances of etcd are actually used: 127.0.0.1:12379 and 127.0.0.1:2379. 

The `balancer` phase in `conf_server` will choose one of them based on the load balancing policy.

If select one of the instances through the `balancer` of `conf_server` when granting lease, and select another instance through the `balancer` of `conf_server` when writing data with lease, the `requested lease not found` error will occur. Because the lease is not issued by the etcd instance that selected to  write the data.