fix(openid-connect): Fix the default 'redirect_uri' (#2426)
Previously the `redirect_uri` was set to `ngx.var.request_uri` if not
configured.  However, it caused the underlying `lua-resty-openidc`
module to raise this error:

```
request to the redirect_uri path but there's no session state found
```

because `lua-resty-openidc` would think it was the redirection response
from OP when the `redirect_uri` equals `ngx.var.request_uri`.

Although the OAuth 2.0 Security Best Current Practice [1] recommends
that the `redirect_uri` should be explicitly specified to prevent
malicious redirection attacks, it would also be handy for APISIX to
properly determine a default one if `redirect_uri` not given.

Therefore, append the `.apisix/redirect` suffix to the current request
URI to determine the default `redirect_uri`.  It makes
`lua-resty-openidc` happy and it's almost unlikely to conflict with
user's URIs.

Also note that the OP should be properly configured to accept such
auto-determined redirect URIs.

Update the documentation accordingly.

Fix #2426.

[1] https://datatracker.ietf.org/doc/draft-ietf-oauth-security-topics/