feat(openid-connect): Add 'symmetric_key' option to support HS??? tokens
The underlying `lua-resty-openidc` module already supports the
`symmetric_key` option to specify the HMAC key for verifying HS???
tokens.  However, note that `lua-resty-openidc` just passes the
`symmetric_key` value as-is to HMAC.  So we choose to accept a
base64url-encoded secret, which is easier to obtain from OP and
configure, and then decode it before passing it to `lua-resty-openidc`.