ci: avoid image build failure by avoid interacting with sum server
When using Go 1.17+ and go.sum is missing in the module, pack will
generate a go.sum file by interacting with sum server, which will
cause image build failure if it failed to do this.

The solution is generating a go.sum and tracking it in the VCS, so
we don't need to generate it in the CI.

In this PR, I also upgrade and generate go.sum for the other images,
to prevent the same thing from happening in the future.
Signed-off-by: spacewander <spacewanderlzx@gmail.com>