Updated default HTTP router
Updated the default HTTP router from radixtree_uri to radixtree_host_uri to remove the confusion that which route will hit first.

Let's say, we have two routes, the first one requires the host matches *.example.com, and the URI path is /anything, And the second one needs the host matches foo.example.com exactly, and the URI path is also /anything. In such a case, if we send an API request, we may hit the first route, which is counterintuitive.