Use mtime instead of ctime
In order to avoid unnecessary reloading of the yaml configuration. The time used to detect the last change to the config file is changed from `change` which is the time of the last file status change to the `modification` time wich is the time of the last data modification.

For more information also checkou http://lunarmodules.github.io/luafilesystem/manual.html#reference and https://linux.die.net/man/2/stat

Fixes: https://github.com/apache/apisix/issues/9051