Initial prototype to test apisix.yaml
This prototype tries to show a way to validate the `apisix.yaml` by starting openresty and only running the validation code.

As part of the validation we initialize the configuration and then check if no errors where logs. If any errors are logged then the `apisix test` command fails.

Special thanks to https://github.com/openresty/resty-cli/!