feat(datadog): add include_method option
Currently there is no way to distinguish Datadog metrics for e.g. GET &
POST requests for the same endpoint within a single Apisix route,
although their performance characteristics are likely to be quite
different.

With these changes, if `include_method` is set to true, HTTP method is
included as a metric tag with the `method:` key, enabling such requests
to be differentiated.