feat(datadog): add response_status_class tag to metrics
Having a separate tag for the HTTP response code class allows useful
grouping in Datadog by e.g. successful requests (2xx), client errors
(4xx), and server errors (5xx). The existing `response_status` already
essentially includes that information, but because of how Datadog works,
one would need to list out each possible value between 400 and 431 to
group all client errors, for example. Having a separate tag for the
class greatly simplifies this.

Generally it may be problematic to add new tags by default without an
opt-in configuration option, because Datadog charges based on [custom
metrics count][1] and additional tags can increase that count. However,
the additional tag is safe to add here, because it is guaranteed not to
increase the custom metric count. That is because the new tag is based
on the already included `response_status` value, which is more granular
than the new `response_status_class` value, so the number of unique tag
combinations will not change.

[1]: https://docs.datadoghq.com/metrics/custom_metrics/